Refactor re-usable code into api-helpers
