+++ test/scenarios/unpin-device-after-release.ts
@@ -6,81 +6,12 @@ import { supertest, UserObjectParam } from '../test-lib/supertest';
 import { version } from '../test-lib/versions';
 
 import * as fixtures from '../test-lib/fixtures';
-
-interface MockReleaseParams {
-	belongs_to__application: number;
-	is_created_by__user: number;
-	commit: string;
-	composition: string;
-	status: string;
-	source: string;
-	build_log: string;
-	start_timestamp: number;
-	end_timestamp?: number;
-	update_timestamp?: number;
-}
-interface MockImageParams {
-	is_a_build_of__service: number;
-	start_timestamp: number;
-	end_timestamp: number;
-	push_timestamp: number;
-	status: string;
-	image_size: number;
-	project_type?: string;
-	error_message?: string;
-	build_log: string;
-}
-interface MockServiceParams {
-	application: number;
-	service_name: string;
-}
-
-type MockRelease = MockReleaseParams & { id: number };
-type MockImage = MockImageParams & { id: number };
-type MockService = MockServiceParams & { id: number };
-
-const addReleaseToApp = async (
-	auth: UserObjectParam,
-	release: MockReleaseParams,
-): Promise<MockRelease> =>
-	(await supertest(auth).post(`/${version}/release`).send(release).expect(201))
-		.body;
-
-const addImageToService = async (
-	auth: UserObjectParam,
-	image: MockImageParams,
-): Promise<MockImage> =>
-	(await supertest(auth).post(`/${version}/image`).send(image).expect(201))
-		.body;
-
-const addServiceToApp = async (
-	auth: UserObjectParam,
-	serviceName: string,
-	application: number,
-): Promise<MockService> =>
-	(
-		await supertest(auth)
-			.post(`/${version}/service`)
-			.send({
-				application,
-				service_name: serviceName,
-			})
-			.expect(201)
-	).body;
-
-const addImageToRelease = async (
-	auth: UserObjectParam,
-	imageId: number,
-	releaseId: number,
-): Promise<void> => {
-	await supertest(auth)
-		.post(`/${version}/image__is_part_of__release`)
-		.send({
-			image: imageId,
-			is_part_of__release: releaseId,
-		})
-		.expect(201);
-};
+import {
+	addReleaseToApp,
+	addImageToService,
+	addServiceToApp,
+	addImageToRelease,
+} from '../test-lib/api-helpers';
 
 describe('Device with missing service installs', () => {
 	let fx: fixtures.Fixtures;
